Ties now give half a point

+++ TicTacToe/Code/Game/Game.hpp
@@ -49,7 +49,7 @@ public:
 
 struct metaData_t
 {
-	int m_numberOfWins = 0;
+	float m_numberOfWins = 0;
 	int m_numberOfSimulations = 0;
 };
